+++ main-service/src/main/java/ru/practicum/ewm/main_service/event/util/SortShortDtoByViewsComparator.java
@@ -9,7 +9,7 @@ public class SortShortDtoByViewsComparator implements Comparator<EventShortDto> 
 
     @Override
     public int compare(EventShortDto e1, EventShortDto e2) {
-        return (int) (e2.getViews() - e1.getViews());
+        return e2.getViews() - e1.getViews();
     }
 
     @Override
